Half way through adding query List results
added proper httpurl connection framework. Setup an async task to
download the query results into. Setup callback interfaces to the
activity so the fragment can retrieve data when it needs it. Build the
url from the data retrieved in the callbacks. Passed the url to the
async task for download and parsing. Still need a list view loader task
to take care of loading the list view async style. Still need a JSON
parser to parse the places JSON.