Geofences being listed added and deleted
Geofences are now being listed on the list view when there is no query
to be displayed. They are also being deleted properly. Not sure why it
is not refreshing the list view when the geofence is deleted or added
but that could probably be solved with a little more testing.